RIP-71, more logging in /dev_auth; front-end layout renaming
