+++ src/layouts/standalone/View.vue
